Moved utility classes from ModelUtils to IdentificationUtils
